requirements.txt
google-generativeai==0.5.2
pandas>=2.2.2
scikit-learn>=1.5.0
numpy==1.26.4
joblib==1.4.2
sentence-transformers==2.7.0
pyyaml==6.0.1
python-dotenv==1.0.1
torch==2.3.0
requests>=2.31.0
reportlab>=4.1.0
streamlit>=1.28.0
